created Reader and Validator Class and moved code
